Code review and refactoring.
- job search API's was moved to a seperate file.
- Submit application was moved to a seperate file.
- job list was moved to a seperate file.
- some functions related to job data formating was removed to a seperate file.
